changed loginForm from JS to TSX, moved it to components folder. Removed optional tag on User type languageIds. Moved login service to /services
